requirements.txt
streamlit>=1.53.0
plotly>=6.0.0
pandas>=2.0.0
numpy>=2.0.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
arch>=7.0.0
ruptures>=1.1.0
scipy>=1.12.0
antropy>=0.1.6
hmmlearn>=0.3.0
yfinance>=1.0.0
fredapi>=0.5.0
matplotlib>=3.8.0
nelson-siegel-svensson>=0.5.0
anthropic>=0.40.0
openai>=1.50.0
fpdf2>=2.7.0
pyarrow>=14.0.0
